Move alignments to `geom` subpkg
