+++ moldesign/geom/__init__.py
@@ -10,3 +10,4 @@ from .constraints import *
 from .symmetry import *
 from .monitor import *
 from .shake import *
+from .alignment import *
