+++ moldesign/mathutils/__init__.py
@@ -1,0 +1,2 @@
+from .vectormath import *
+from .eigen import *
