+++ moldesign/mathutils/eigen.py
@@ -1,7 +1,7 @@
 from __future__ import print_function, absolute_import, division
-from future.builtins import *
 from future import standard_library
 standard_library.install_aliases()
+from future.builtins import *
 
 # Copyright 2017 Autodesk Inc.
 #
@@ -16,9 +16,7 @@ standard_library.install_aliases()
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import operator as op
 
-import numpy as np
 from .. import units as u
 from ..utils import exports
 from ..mathutils import normalized
@@ -26,7 +24,7 @@ from ..mathutils import normalized
 
 @exports
 class Eigenspace(object):
-    """ Holds the eigenvactors and eigenvalues of a given matrix.
+    """ Holds sets of eigenvactors and eigenvalues, offers helpful methods for organizing them.
 
     Args:
         evals (List[Scalar]): list of eigenvalues
@@ -37,10 +35,10 @@ class Eigenspace(object):
         The eigenvectors from many eigenvector solvers - notably including scipy's - will need to be
         transposed to fit the form of ``evecs`` here!
 
-        It's generally assumed that Eigen space objects will be subclassed to wrap various bits of
+        It's generally assumed that Eigenspace objects will be subclassed to wrap various bits of
         eignproblem-related functionality.
     """
-    def __init__(self, evals, evecs, real=True):
+    def __init__(self, evals, evecs):
         if len(evals) != len(evecs):
             raise ValueError('Eigenvalues and eigenvectors have different lengths!')
 
@@ -100,81 +98,3 @@ class Eigenspace(object):
             raise ValueError('Transform accepts only 1- or 2-dimensional arrays')
 
 
-@exports
-class PrincipalMomentsOfInertia(Eigenspace):
-    """ Calculates the moments of inertia of a molecule at a given position
-
-    Notes:
-        This object is NOT updated if the initializing molecule's positions change.
-        This allows it to be used to reorient other molecules into the same coordinate system
-
-    Args:
-        mol (moldesign.Molecule): molecule to calculate PMIs for
-        mass_centered (bool): calculate PMIs relative to molecule's center of mass (default: True)
-
-    Attributes:
-        axes (List[Vector[len=3]]]): principal rotation axes
-        moments (Vector[length**2 * mass, len=3]): rotational inertias for the three axes
-    """
-    def __init__(self, mol, mass_centered=True):
-        from scipy.linalg import eig
-        inertia_tensor, com = get_inertia_tensor(mol, mass_centered=mass_centered)
-
-        evals, evecs = eig(inertia_tensor.defunits_value())  # strip units before going to scipy
-        evals = evals*u.default.mass*u.default.length ** 2  # reapply units to evals
-
-        assert (abs(evecs.imag) < 1e-12).all()  # hopefully everything is real?
-        assert (abs(evals.imag) < 1e-12 * evals.units).all()
-
-        super().__init__(evals.real, evecs.T)
-        self._com = com
-        self.sort()
-
-    @property
-    def moments(self):
-        return self.evals
-
-    @property
-    def axes(self):
-        return self.evecs
-
-    def reorient(self, mol):
-        """ Rotates molecular coordinates into the PMI coordinate system
-
-        The molecule will be translated (if necessary), then rotated so that x lies along PMI 1,
-        y along PMI 2, and z along PMI 3.
-
-        Args:
-            mol (moldesign.Molecule): molecule to re-orient - its current positinos will be changed
-        """
-        if self._com is not None:
-            mol.com -= self._com
-        mol.positions = self.transform(mol.positions)
-
-
-@exports
-def get_inertia_tensor(mol, mass_centered=True):
-    """ Calculates the moment of inertia for a molecule
-
-    Args:
-        mol (moldesign.Molecule): calculate the inertia tensor of this  molecule
-        mass_centered (bool): if True, calcualte tensor relative to molecule's center of mass. If
-           False, calculate it relative to the current origin.
-
-    Returns:
-        Matrix[mass*length*length]: moment of inertia tensor
-    """
-    inertia_tensor = np.zeros((3,3)) * u.default.mass * u.default.length**2
-    com = mol.com if mass_centered else None
-
-    pos = mol.positions
-    if mass_centered:
-        pos = pos - mol.com
-
-    for i in range(3):
-        inertia_tensor[i,i] = (mol.masses * (pos[:,i-1]**2 + pos[:,i-2]**2)).sum()
-        for j in range(i+1,3):
-            di = (mol.masses * pos[:,i] * pos[:,j]).sum()
-            inertia_tensor[i,j] = -di
-            inertia_tensor[j,i] = -di
-    return inertia_tensor, com
