+++ moldesign/tools/__init__.py
@@ -5,4 +5,3 @@ __all__ = []
 
 from .topology import *
 from .build import *
-from .eigen import *
